fix: move isConfirmWord to utils to satisfy react-refresh lint rule
